+++ NEWS.md
@@ -2,6 +2,7 @@
 * Made change to MSARmdl(), MSVARmdl(), and HMmdl(). Specifically, when msmu or msvar is FALSE, output list copies single regime value for each k regime. This is needed for simulating the null when either msmu or msvar is FALSE in LMCLRTest() and MMCLRTest().
 * Added option to use different number of initial values for estimating MSMs with observed data vs. for null distribution (see documentation for LMCLRTest() and MMCLRTest()).
 * Updated USGNP data set to include 2022 Q3.
+* MMCLRTest() now has the option to add lower and upper bounds for autoregressive coefficients to help reduce errors from polyroot() when optimizing. MMC_bounds() has been updated to reflect this. 
 
 # MSTest 0.1.1